@@ -14,34 +14,46 @@
             "email": "saschaprolic@googlemail.com"
         }
     ],
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^7.1",
+        "ext-mongodb": "^1.5.2",
+        "mongodb/mongodb": "^1.4.2",
         "prooph/event-store": "^6.3",
         "beberlei/assert": "~2.4",
         "prooph/common": "^3.7"
     },
     "require-dev": {
-        "container-interop/container-interop": "^1.1",
-        "sandrokeil/interop-config": "^1.0",
-        "phpunit/phpunit": "4.8.* || ^5.3",
-        "sebastian/comparator": "^1.2",
-        "fabpot/php-cs-fixer": "1.7.*",
+        "malukenho/docheader": "^0.1.4",
+        "phpspec/prophecy": "^1.7",
+        "phpunit/php-invoker": "^2.0",
+        "phpunit/phpunit": "^7.1.4",
+        "prooph/php-cs-fixer-config": "^0.2.2",
+        "psr/container": "^1.0",
+        "sandrokeil/interop-config": "^2.0.1",
         "satooshi/php-coveralls": "^1.0"
     },
     "suggest": {
-        "container-interop/container-interop": "For usage of provided factories",
-        "sandrokeil/interop-config": "For usage of provided factories",
-        "alcaeus/mongo-php-adapter": "Allows usage of PHP 7",
-        "ext-mongo": ">=1.6 for usage with mongo extension and PHP 5.x",
-        "ext-mongodb": ">=1.1 for usage with mongodb extension and PHP 7"
+        "psr/container": "For usage of provided factories",
+        "sandrokeil/interop-config": "For usage of provided factories"
     },
     "config": {
         "preferred-install": {
             "prooph/*": "source"
-        }
+        },
+        "sort-packages": true
     },
-    "conflict": {
-        "sandrokeil/interop-config": "<1.0"
+    "scripts": {
+        "check": [
+            "@cs",
+            "@docheader",
+            "@test"
+        ],
+        "docheader": "vendor/bin/docheader check examples/ src/ tests/",
+        "cs": "php-cs-fixer fix -v --diff --dry-run",
+        "cs-fix": "php-cs-fixer fix -v --diff",
+        "test": "phpunit"
     },
     "autoload": {
         "psr-4": {
